Added requirement that git must be installed
When git is not installed, the initial bootstrap script fails, but the rest of the script continues and the errors don't obviously point to git not being installed. A "try/catch" around the some of the commands would be best, but for now just clarifying that git should be installed is good enough, I think.